requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.4
dataclasses-json==0.6.7
distro==1.9.0
faiss-cpu==1.13.2
filetype==1.2.0
frozenlist==1.8.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.189.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-genai==1.62.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
greenlet==3.3.1
grpcio==1.78.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.43
langchain-google-genai==1.0.8
langchain-text-splitters==0.2.4
langsmith==0.1.147
marshmallow==3.26.2
multidict==6.7.1
mypy_extensions==1.1.0
numpy==1.26.4
orjson==3.11.7
packaging==23.2
propcache==0.4.1
proto-plus==1.27.1
protobuf==4.25.8
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
pypdf==6.6.2
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.46
tenacity==8.5.0
tqdm==4.67.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.0
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
